eva_annotation_board_backend/pyproject.toml
[project]
name = "eva-annotation-board-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Dain Park",email = "zndgn555@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-admin-rangefilter (>=0.13.2,<0.14.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
